Primjena happy hour popusta. Ostali sitni bugfixevi.

+++ RP3_projekt/RP3_projekt/Bill.cs
@@ -12,8 +12,8 @@ namespace RP3_projekt
         public DateTime Time { get; set; }
         public Employee Employee { get; set; }
         public List<Item> Items { get; set; }
-        public double Price { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal TotalPrice { get; set; }
         public List<Discount> Discounts { get; set; }
     }
 }
